Added command line interface in Main class. Made it easier to merge BestFirstSearch class with the Main class.
Added a command line interface in the Main class. The user can pick their origin, destination, and choice of search algorithm from a preset example.

Changed the output of the UniformCostSearch class to an ArrayList of Node objects. The expected output is a list of node objects in path order.